+++ work_queue/src/work_queue_worker.c
@@ -1621,20 +1621,20 @@ If 0, the worker is using more resources than promised. 1 if resource usage hold
 static int enforce_worker_limits(struct link *manager)
 {
 	if( manual_disk_option > 0 && local_resources->disk.inuse > manual_disk_option ) {
-		fprintf(stderr,"work_queue_worker: %s used more than declared disk space (--disk - < disk used) %"PRIu64" < %"PRIu64" MB\n", workspace, manual_disk_option, local_resources->disk.inuse);
+		fprintf(stderr,"work_queue_worker: %s used more than declared disk space (--disk - < disk used) %"PRIu64" < %lf MB\n", workspace, manual_disk_option, local_resources->disk.inuse);
 
 		if(manager) {
-			send_manager_message(manager, "info disk_exhausted %lld\n", (long long) local_resources->disk.inuse);
+			send_manager_message(manager, "info disk_exhausted %lf\n", local_resources->disk.inuse);
 		}
 
 		return 0;
 	}
 
 	if(manual_memory_option > 0 && local_resources->memory.inuse > manual_memory_option) {
-		fprintf(stderr,"work_queue_worker: used more than declared memory (--memory < memory used) %"PRIu64" < %"PRIu64" MB\n", manual_memory_option, local_resources->memory.inuse);
+		fprintf(stderr,"work_queue_worker: used more than declared memory (--memory < memory used) %"PRIu64" < %lf MB\n", manual_memory_option, local_resources->memory.inuse);
 
 		if(manager) {
-			send_manager_message(manager, "info memory_exhausted %lld\n", (long long) local_resources->memory.inuse);
+			send_manager_message(manager, "info memory_exhausted %lf\n", local_resources->memory.inuse);
 		}
 
 		return 0;
@@ -1658,10 +1658,10 @@ static int enforce_worker_promises(struct link *manager)
 	}
 
 	if( manual_disk_option > 0 && local_resources->disk.total < manual_disk_option) {
-		fprintf(stderr,"work_queue_worker: has less than the promised disk space (--disk > disk total) %"PRIu64" < %"PRIu64" MB\n", manual_disk_option, local_resources->disk.total);
+		fprintf(stderr,"work_queue_worker: has less than the promised disk space (--disk > disk total) %"PRIu64" < %lf MB\n", manual_disk_option, local_resources->disk.total);
 
 		if(manager) {
-			send_manager_message(manager, "info disk_error %lld\n", (long long) local_resources->disk.total);
+			send_manager_message(manager, "info disk_error %lf\n", local_resources->disk.total);
 		}
 
 		return 0;
@@ -2898,7 +2898,7 @@ int main(int argc, char *argv[])
 	connect_stoptime = time(0) + connect_timeout;
 
 	measure_worker_resources();
-	printf("work_queue_worker: using %"PRId64 " cores, %"PRId64 " MB memory, %"PRId64 " MB disk, %"PRId64 " gpus\n",
+	printf("work_queue_worker: using %lf cores, %lf MB memory, %lf MB disk, %lf gpus\n",
 		total_resources->cores.total,
 		total_resources->memory.total,
 		total_resources->disk.total,
